Rename base_spec_helper to spec_helper
