+++ spec/helpers/users_helper_spec.rb
@@ -1,4 +1,4 @@
-require "base_spec_helper"
+require "spec_helper"
 require "./app/helpers/users_helper"
 
 describe UsersHelper do
